Update newrelic.md
The FutureStack 2022 event held in April marked the debut of New Relic’s new brand - complete with a new look and feel including the company logo. Updated this documentation with the latest logo and the link to Algorithmia quickstart page for  better user experience of Algorithmia users. Please check and approve.